requirements.txt
flask
flask_sqlalchemy
sqlalchemy
pyodbc
selenium
pytest
requests
pyyaml
pytest-html
PyJWT
allure-pytest
pytest-order